Fix deployment
We needed a way to notify gulp that generate is done only when every
file gets written on disk. To do this, have PagesWriter inherit from
EventEmitter, then emit a 'write' event when every render call is
completed.

- Remove the public folder before deployment.
- Change PagesWriter to not take config on instantiation, only on
  setState. This way only calling setState can ever trigger a write
  instead of having two ways to trigger a write.
- Changed paths to return an array instead of set so that we can map and
  pass it to Promise.all more easily.

I had thought about making setState return a promise instead, but since
we are debouncing writes, it was difficult to resolve the promise
correctly because the write may not be called.