Switch to browserify instead of webpack
Use the `bundleExternal: false` feature of browserify to make our
renderer module smaller. The renderer will require third party modules
like react and react-router, and they should all be able to be ran in
the node environment, so no need to bundle them into the module. In
other words, we're only bundling up our renderer module for the ES6
transformation step in browserify.

Also added a separate gulp task for js watch and build. The dependency
of the generate task is js:build in production or js:watch in
development. This fixes a problem where the deploy script would hang if
we were using js:watch instead of js:build.

Also updated the about page.