Update Index component when post is updated or deleted
The Index page shows post excerpts so the markup may be updated if a
post's first paragraph is updated. The index page may also change if a
post is deleted. Therefore, post updates should update the Index page.

A simple way to achieve this is to rebuild the entire site on a post
update.

The issue is this slows down the live reload if we're in the post page
of the post that was just updated.

To get around this slow down, write the page for the post first, then
rebuild the site. In addition to that, add a check in writeReactElement
to cache the content by the url. This way, we don't rewrite out a file
that isn't updated.