More efficient clearing of the require cache
Hook into the node's `require` to construct a dependency graph. When a
file is saved and the file is in the dependency graph, clear the cache
of all of the dependants.

Previously we were just naively removing any modules not in
`node_modules`. This change should make it more efficient.