Remove farmhash
Farmhash is causing issues in gitlab CI; it won't build.

Instead of hashing the title to produce the color of the post, just make
the function return a new color every time it is called. However, the
result is cached so that the same input returns the same output per
process.