ssg: return promise from build
build() now returns a promise to signify that the build process has
completed. it will return an array of rows. if a row was deleted
since a previous build, it will have the `deleted: true` flag, otherwise
it will have the `data` field which is what was written to disk.

this also required moving the warning of the row shape to the manifest
generation instead of the build step to make invalid rows easier to
remove and diff.